Refactored calculation state
