+++ webapp/loggedin/modules/analysis/calculation/state/calculationState.js
